pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "repl-mcp"
version = "0.2.0"
description = "A TUI application and MCP server for managing interactive REPL programs"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "textual",
    "mcp[cli]",
    "uvicorn",
    "pyte",
    "unique-namer",
]

[project.optional-dependencies]
dev = [
    "black",
    "pylint",
    "pytest",
    "pytest-asyncio",
    "pre-commit",
]

[project.scripts]
repl-mcp = "repl_mcp.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pylint.format]
max-line-length = 100

[tool.pylint."messages control"]
disable = ["C0114", "C0115", "C0116"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
